requirements/testing.txt
-r base.txt
django-mailer==0.1.0
factory-boy==2.4.1
django-libs==1.64.7
selenium==2.44.0
